+++ src/app/components/package/package-list/package-list.component.ts
@@ -3,9 +3,10 @@ import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { PackageFilter } from '../package-filter';
 import { PackageService } from '../package.service';
 import { Package } from '../package';
-import { MatSort, MatPaginator, MatTableDataSource } from '@angular/material';
+import { MatSort, MatPaginator, MatTableDataSource, MatDialog } from '@angular/material';
 import { Router } from '@angular/router';
 import { tap, startWith } from 'rxjs/operators';
+import { WarningDialog } from '../../warning-dialog/dialog-warning';
 
 @Component({
     selector: 'package',
@@ -36,7 +37,7 @@ export class PackageListComponent implements OnInit, AfterViewInit {
     dataSource = new MatTableDataSource<any>([]);
 
     displayedColumns: string[] = ['No', 'name', 'description', 'action']
-    constructor(private packageService: PackageService, private router: Router) {
+    constructor(private packageService: PackageService, private dialog: MatDialog) {
     }
 
     ngOnInit() {
@@ -71,13 +72,21 @@ export class PackageListComponent implements OnInit, AfterViewInit {
             )
     }
 
-    removePackage(pack, index) {
-        this.packageService.delete(pack).subscribe((response: any) => {
-            console.log(response)
-            if (response.status === 200) {
-                this.getPackageList(this.paginator.pageIndex, this.paginator.pageSize)
+    removePackage(pack) {
+
+        this.dialog.open(WarningDialog, {
+            width: '400px',
+            data: { title: 'Warning', message: `Are you sure want to delete ${pack.name} Package` }
+        }).afterClosed().subscribe((result) => {
+            if (result) {
+                this.packageService.delete(pack).subscribe((response: any) => {
+                    if (response.status === 200) {
+                        this.getPackageList(this.paginator.pageIndex, this.paginator.pageSize)
+                    }
+                }, error => console.log(error));
             }
-        })
+        });
+
     }
 
     getPackageCount() {
